Implement graph and pathfinding algorithms
This branch adds the script that was used to generate the data set. It
also adds some boilerplating for parsing the JSON data set on a
nonblocking thread. The bulk of this branch however is dedicated to the
Graph class, which parses the JSON data set object in memory into an
Graph object, which is just a wrapper around an adjacency list with some
methods. This class also has public functions for the different path
finding algorithms. Thus far, DFS, BFS, Dijkstra's, and A-Star have been
implemented.